Make specs more explicit by removing fixture
