+++ spec/ostruct/sanitizer_spec.rb
@@ -18,6 +18,12 @@ describe OStruct::Sanitizer do
   end
 
   describe "#truncate" do
+    class User < OpenStruct
+      include OStruct::Sanitizer
+      truncate :first_name, :last_name, length: 10
+      truncate :middle_name, length: 3, strip_whitespaces: false
+    end
+
     let(:user) do
       User.new(
         first_name: " first name longer than 10 characters",
@@ -46,6 +52,11 @@ describe OStruct::Sanitizer do
   end
 
   describe "#drop_punctuation" do
+    class User < OpenStruct
+      include OStruct::Sanitizer
+      drop_punctuation :city, :country
+    end
+
     let(:user) do
       User.new(
         city: "Porto, Alegre!",
@@ -69,6 +80,11 @@ describe OStruct::Sanitizer do
   end
 
   describe "#strip" do
+    class User < OpenStruct
+      include OStruct::Sanitizer
+      strip :email, :phone
+    end
+
     let(:user) do
       User.new(
         email: "  drborges.cic@gmail.com   ",
@@ -92,6 +108,11 @@ describe OStruct::Sanitizer do
   end
 
   describe "#digits" do
+    class User < OpenStruct
+      include OStruct::Sanitizer
+      digits :ssn, :cell_phone
+    end
+
     let(:user) do
       User.new(
         ssn: "111-11-1111",
