+++ spec/spec_helper.rb
@@ -1,5 +1,2 @@
 $LOAD_PATH.unshift File.expand_path("../../lib", __FILE__)
 require "ostruct/sanitizer"
-Dir['spec/fixtures/*.rb'].each do |f|
-  require_relative "../#{f}"
-end
